+++ src/main/java/app/onepass/organizer/repositories/TagRepository.java
@@ -6,4 +6,5 @@ import org.springframework.stereotype.Repository;
 import app.onepass.organizer.entities.TagEntity;
 
 @Repository
-public interface TagRepository extends JpaRepository<TagEntity, Long> {}
+public interface TagRepository extends JpaRepository<TagEntity, Integer> {
+}
